feat: 프로젝트/직원 API 응답 보강 및 기초 데이터 생성 API 추가

+++ src/main/java/com/example/dbteam2backend/project/dto/AssignmentRequestDto.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
 @NoArgsConstructor
 public class AssignmentRequestDto {
     private Long projectId;
-    private Long developerId;
+    private Integer developerId;
     private Long roleId;
     private LocalDate startDate;
     private LocalDate endDate;
